# CommandLineParser improvements