--- a/build.sbt
+++ b/build.sbt
@@ -568,7 +568,7 @@ lazy val partest = configureAsSubproject(project)
   )
 
 lazy val bench = project.in(file("test") / "benchmarks")
-  .dependsOn(library)
+  .dependsOn(library, compiler)
   .settings(instanceSettings)
   .settings(disableDocs)
   .settings(disablePublishing)