Tokenizing used to snip the line and knit the args,
this commit denits the deknitting.

Adds a benchmark showing linear behavior and a unit test
to show correctness.

Departing from the previous notion of correctness,
internal quotes are respected. `"abc"xyz` is `abcxyz`.

"It is imperative that this be fixed!"