@SethTisue 2018-05-11
rebasing on #6624 will fix the Travis-CI failure.

but also, Scabot isn't triggering Jenkins jobs, we'll look into that.