@SethTisue 2018-05-11
I took the liberty of rebasing, hoping it would reset Scabot & Jenkins