@som-snytt 2018-05-11
I'm going to stop checking my phone for jmh benchmark notifications.

Simply quoted args are surprisingly expensive; embedded quotes like `"abc"xyz` require a bit more work. Wait, I keep comparing different factors of ten.

```
Before:
[info] Benchmark                                    (argCount)  Mode  Cnt        Score       Error  Units
[info] CommandLineParserBenchmark.parsingBenchmark         100  avgt   20    22865.907 ±   402.525  ns/op
[info] CommandLineParserBenchmark.parsingBenchmark         200  avgt   20    76535.892 ±   312.987  ns/op
[info] CommandLineParserBenchmark.parsingBenchmark         500  avgt   20   391772.920 ±  4060.599  ns/op
[info] CommandLineParserBenchmark.parsingBenchmark        1000  avgt   20  1456112.252 ± 15976.252  ns/op
 
After:
[info] Benchmark                                                 (argCount)  Mode  Cnt         Score         Error  Units
[info] CommandLineParserBenchmark.embeddedQuoteParsingBenchmark        1000  avgt   20    113206.128 ±   14776.228  ns/op
[info] CommandLineParserBenchmark.embeddedQuoteParsingBenchmark       10000  avgt   20   1230762.947 ±    9494.246  ns/op
[info] CommandLineParserBenchmark.embeddedQuoteParsingBenchmark      100000  avgt   20  15673465.207 ± 2398766.387  ns/op
[info] CommandLineParserBenchmark.parsingBenchmark                     1000  avgt   20     40582.655 ±    3305.774  ns/op
[info] CommandLineParserBenchmark.parsingBenchmark                    10000  avgt   20    389045.591 ±   41466.469  ns/op
[info] CommandLineParserBenchmark.parsingBenchmark                   100000  avgt   20   5086895.949 ±  506776.209  ns/op
[info] CommandLineParserBenchmark.quoteUnquoteParsingBenchmark         1000  avgt   20     86258.217 ±   25262.772  ns/op
[info] CommandLineParserBenchmark.quoteUnquoteParsingBenchmark        10000  avgt   20    906925.002 ±  151254.491  ns/op
[info] CommandLineParserBenchmark.quoteUnquoteParsingBenchmark       100000  avgt   20   8894730.967 ±  437566.873  ns/op

```