@som-snytt 2018-05-16
I still had some self-deprecating puns to make about this contribution, but I'll include them in a follow-up.